backend/requirements.txt
# Backend Dependencies
fastapi==0.115.2
uvicorn==0.23.2
openai==1.22.0
python-pptx==1.0.2
python-dotenv==1.0.1
tenacity==8.2.3
pydantic==2.6.4

# Test Dependencies (for backend tests)
httpx==0.27.0
pytest>=7.0.0
pytest-mock>=3.10.0
